:recycle: - refactor by splitting out boid-specific logic
